タッチ判定追加
