+++ Assets/Script/Scene/GameMain.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fc33b53d1e91e4fbd938d17b76ca42fd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
