+++ Assets/Script/Scene/GameMain/GameMainManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f09b56625f5c1479cb0237a816156779
+guid: 45f1415a8c4b34dff93fc114f4ea33f1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
